requirements.txt
uvicorn~=0.17.4
python-dotenv~=0.19.2
fastapi~=0.73.0
pydantic~=1.9.0